requirements.txt
playwright==1.49.0
requests>=2.31.0
playwright-stealth>=1.0.6
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
aiohttp>=3.8.0
pandas>=1.5.0
tqdm>=4.65.0
python-dotenv>=1.0.0
openpyxl>=3.10.0
paramiko>=3.0.0
# anthropic>=0.18.0  # Removed - LLM calls now handled by Celigo
